scraper/requirements.txt
fotmob-wrapper
aiohttp
requests
pandas
soccerdata
lxml